try alternative sql statements
This commit contains an alternative "insert" statement, using merge without on match clause (thus making it a conditional insert).
It should be semantically identical to the original (i think, but it's getting late...)

The plan is to test this statement vs the original in a stress test.

ps: there is still no PK on the table. it might be, that the merge command benefits from a PK. I should test this...